Fix QString::arg() usage in chat
It is bad practice to chain .arg(a).arg(b) calls because 'a' might
contain a format specifier and produce unexpected results.  Instead use
the QString::arg() function that takes multiple arguments in a single
pass.  This eliminates the problem with injecting format specifiers.

Also fix the topic setting logic, which was inverted.

Reported-by: Ikkei Shimomura <ikkei.shimomura@gmail.com>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>