Exit program if the Connect... dialog is cancelled
It's not safe to return from the QMainWindow constructor before our
member fields are set up.  Instead use exit(3) for now so that the
program stops and our destructor is not invoked.

In the future it should be possible to display the main window while not
connected to a server.  The Connect... dialog will simply be a step that
automatically is taken when the program starts or that can be manually
performed by the user from a File | Connect... menu item.  But for now
keep the big ugly QMainWindow constructor.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>