Provide a working directory for the client
The client needs a working directory (alsa known as the "session
directory") where it stores .ogg files.  This directory can be preserved
but it's often handy to automatically delete it when disconnecting from
the server.  (Normally users who wish to preserve it will remix the jam
session later.)

This patch is the final piece that allows qtclient to play back remote
audio!

Future patches will add a QSettings variable so that the working
directory path can be customized.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>