Remove Jesusonic DLL support
The Jesusonic DLL provided real-time effects on local channels.  It was
a proprietary Windows DLL that contained a custom programming language
suitable for implementing audio effects.  Jesusonic was never
open-sourced and is integrated into Reaper today without any further
stand-alone DLL development.

I doubt anyone uses the Jesusonic DLL and do not plan on supporting it
in the Qt client.  Furthermore the licensing situation is questionable.

If real-time effects are desirable we should implement these via JACK,
LADSPA, and/or VST in the future instead.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>