Add remote channels to ChannelTreeWidget
This patch adds remote audio channels to the channel tree, making
it possible to see users that are connected to the server and to toggle
mute.

The NJClient has a flag that indicates when remote user information has
changed.  This patch adds the userInfoChanged() signal to
ClientRunThread and implements a slot in MainWindow to enumerate remote
channels each time remote user information changes.

Since we don't know exactly what changed we use a transaction mechanism
which prunes unused audio channels and users when commit() is called.
This can happen when a remote user disconnects from the server.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>