Move audio cleanup to MainWindow::Disconnect()
This patch is analogous to commit a429950b where the
MainWindow::Connect() function was created.  This time we need to split
out the client disconnect and audio cleanup code.

The point of separating MainWindow lifecycle from Connect()/Disconnect()
is that the application can continue running while we connect and
reconnect to servers multiple times.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>